backendPostgresql/polucion_aire_api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyjwt==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
cryptography==41.0.3
python-telegram-bot==20.5
SQLAlchemy==2.0.20
marshmallow==3.20.1
inject==4.3.1
python-jose==3.3.0
gunicorn==21.2.0